feat: add buy list and purchased list to db

+++ CA6/back/src/main/java/services/BuyListService.java
@@ -4,6 +4,7 @@ import entities.BuyList;
 import entities.Commodity;
 import entities.CommodityUserId;
 import entities.User;
+import exceptions.AlreadyInBuyList;
 import exceptions.CommodityIsNotInBuyList;
 import org.springframework.stereotype.Service;
 import repositories.BuyListRepository;
@@ -19,9 +20,18 @@ public class BuyListService {
         this.buyListRepository = buyListRepository;
     }
 
-    public void addItem(Commodity commodity, User user) {
-        BuyList buyList = new BuyList(commodity, user);
-        buyListRepository.save(buyList);
+    public void addItem(Commodity commodity, User user) throws AlreadyInBuyList {
+        CommodityUserId id = new CommodityUserId(commodity.getId(), user.getUsername());
+        Optional<BuyList> buyListOptional = buyListRepository.findById(id);
+
+        if (buyListOptional.isPresent()) {
+            buyListOptional.get().increaseQuantity();
+            buyListRepository.save(buyListOptional.get());
+        } else {
+            BuyList buyList = new BuyList(commodity, user);
+            buyList.setQuantity(1);
+            buyListRepository.save(buyList);
+        }
     }
 
     public void removeItem(String commodityId, String username) throws CommodityIsNotInBuyList {
@@ -29,13 +39,30 @@ public class BuyListService {
         Optional<BuyList> buyListOptional = buyListRepository.findById(id);
 
         if (buyListOptional.isPresent()) {
-            buyListRepository.deleteById(id);
+            if (buyListOptional.get().getQuantity() > 1) {
+                buyListOptional.get().decreaseQuantity();
+                buyListRepository.save(buyListOptional.get());
+            } else
+                buyListRepository.deleteById(id);
         } else {
             throw new CommodityIsNotInBuyList();
         }
     }
 
+    public void deleteItem(CommodityUserId id) {
+        buyListRepository.deleteById(id);
+    }
+
     public List<BuyList> getUserItems(User user) {
         return buyListRepository.findByUser(user);
     }
+
+    public float getCurrentBuyListPrice(String username) {
+        Float totalPriceSum = buyListRepository.getTotalPriceSum(username);
+        if (totalPriceSum == null) {
+            return 0;
+        }
+        return totalPriceSum;
+    }
+
 }
